requirements.txt
absl-py==0.15.0
actionlib==1.14.0
addict==2.4.0
aliyun-python-sdk-core==2.14.0
aliyun-python-sdk-kms==2.16.2
astunparse==1.6.3
bondpy==1.8.6
Brotli @ file:///home/conda/feedstock_root/build_artifacts/brotli-split_1648883617327/work
cachetools==5.3.2
catkin==0.8.10
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1707022139797/work/certifi
cffi==1.15.1
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1698833585322/work
click==8.1.7
colorama==0.4.6
crcmod==1.7
cryptography==42.0.2
cv_bridge==1.13.0
cycler==0.11.0
dynamic-reconfigure==1.7.3
empy==3.3.4
flatbuffers==1.12
fonttools==4.38.0
gast==0.3.3
gencpp==0.7.0
geneus==3.0.0
genlisp==0.4.18
genmsg==0.6.0
gennodejs==2.0.2
genpy==0.6.15
google-auth==2.27.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1701026962277/work
importlib-metadata==6.7.0
jarvis @ file:///home/jihwan98/Anchor3DLane/wheels/jarvis-2021.4.2-py2.py3-none-any.whl
jmespath==0.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
Markdown==3.4.4
markdown-it-py==2.2.0
MarkupSafe==2.1.5
matplotlib==3.5.1
mdurl==0.1.2
message-filters==1.16.0
mmcv-full==1.6.2
-e git+ssh://git@github.com/artzha/Anchor3DLane.git@86fcfbf6518700fedd83ee3c6e73ddd507e6bcd4#egg=mmsegmentation
model-index==0.1.11
munkres==1.1.4
nose @ file:///opt/conda/conda-bld/nose_1642704612149/work
numpy @ file:///home/conda/feedstock_root/build_artifacts/numpy_1649806299270/work
oauthlib==3.2.2
opencv-python==4.1.2.30
opendatalab==0.0.10
openmim==0.3.9
openxlab==0.0.10
opt-einsum==3.3.0
ordered-set==4.1.0
ortools==9.3.10497
oss2==2.17.0
packaging==21.3
pandas==1.3.5
Pillow==9.5.0
platformdirs==4.0.0
prettytable==3.4.1
protobuf==3.20.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.20.0
Pygments==2.17.2
pyparsing==3.1.1
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1648857264451/work
python-dateutil==2.8.2
pytz==2023.4
PyYAML==6.0.1
requests==2.28.2
requests-oauthlib==1.3.1
rich==13.7.0
ros_numpy==0.0.5
rosapi==0.11.16
rosbag==1.16.0
rosboost-cfg==1.15.8
rosbridge-library==0.11.16
rosbridge-server==0.11.16
rosclean==1.15.8
roscreate==1.15.8
rosgraph==1.16.0
roslaunch==1.16.0
roslib==1.15.8
roslz4==1.16.0
rosmake==1.15.8
rosmaster==1.16.0
rosmsg==1.16.0
rosnode==1.16.0
rosparam==1.16.0
rospy==1.16.0
rosservice==1.16.0
rostest==1.16.0
rostopic==1.16.0
rosunit==1.15.8
roswtf==1.16.0
rsa==4.9
scipy==1.4.1
sensor-msgs==1.13.1
Shapely==1.8.4
six==1.15.0
smclib==1.8.6
tabulate==0.8.10
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.4.0
tensorflow-estimator==2.4.0
termcolor==1.1.0
tf==1.13.2
tf2-py==0.7.6
tf2-ros==0.7.6
tomli==2.0.1
topic-tools==1.16.0
torch==1.12.1
torchaudio==0.12.1
torchvision==0.13.1
tqdm==4.64.0
typing-extensions==3.7.4
ujson==5.5.0
urllib3==1.26.18
wcwidth==0.2.13
Werkzeug==2.2.3
wrapt==1.12.1
yapf==0.40.2
zipp==3.15.0
